WPF-956968-Chart KB Sample Improvement
