--- LegendIconAppearanceCustomization/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Syncfusion.Licensing" version="18.1.0.59" targetFramework="net46" />
-  <package id="Syncfusion.SfChart.WPF" version="18.1.0.59" targetFramework="net46" />
-</packages>
