+++ LegendIconCustomization/LegendIconCustomization.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.12.35527.113 d17.12
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LegendIconCustomization", "LegendIconCustomization.csproj", "{F83D144E-8C1A-49F6-8B9A-35368023B74F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{F83D144E-8C1A-49F6-8B9A-35368023B74F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F83D144E-8C1A-49F6-8B9A-35368023B74F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F83D144E-8C1A-49F6-8B9A-35368023B74F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F83D144E-8C1A-49F6-8B9A-35368023B74F}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
