+++ LegendIconCustomization/MainWindow.xaml
@@ -1,0 +1,38 @@
+﻿<Window x:Class="LegendIconCustomization.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:chart="clr-namespace:Syncfusion.UI.Xaml.Charts;assembly=Syncfusion.SfChart.WPF"
+        xmlns:local="clr-namespace:LegendIconCustomization" 
+        mc:Ignorable="d"
+        Title="MainWindow" Height="450" Width="800">
+    <Grid>
+        <Grid.DataContext>
+            <local:ViewModel/>
+        </Grid.DataContext>
+        
+        <chart:SfChart Margin="10">
+            
+            <chart:SfChart.Legend>
+                <chart:ChartLegend></chart:ChartLegend>
+            </chart:SfChart.Legend>
+            
+            <chart:SfChart.PrimaryAxis>
+                <chart:CategoryAxis  LabelFormat="MMM/dd"/>
+            </chart:SfChart.PrimaryAxis>
+            
+            <chart:SfChart.SecondaryAxis>
+                <chart:NumericalAxis />
+            </chart:SfChart.SecondaryAxis>
+            
+            <chart:FastLineSeries Label="Series 1" StrokeDashArray="1,1" ItemsSource="{Binding DataCollection}" XBindingPath="Date" YBindingPath="Value">
+                <chart:FastLineSeries.LegendIconTemplate>
+                    <DataTemplate >
+                        <Polyline Points="0,8,25,8" Stroke="{Binding Interior}" StrokeThickness="{Binding StrokeThickness}" StrokeDashArray="1,1"/>
+                    </DataTemplate>
+                </chart:FastLineSeries.LegendIconTemplate>
+            </chart:FastLineSeries>
+        </chart:SfChart>
+    </Grid>
+</Window>
