+++ LegendIconCustomization/Model.cs
@@ -1,0 +1,14 @@
+﻿namespace LegendIconCustomization
+{
+    public class Data
+    {
+        public DateTime Date { get; set; }
+        public double Value { get; set; }
+
+        public Data(DateTime date, double value)
+        {
+            Date = date;
+            Value = value;
+        }
+    }
+}
